[Fix] Change to Float from Int for Liter data in Activity. Update related for this fix

+++ app/src/main/java/id/vee/android/data/VeeRepository.kt
@@ -238,7 +238,7 @@ class VeeRepository(
         token: String,
         date: String,
         distance: Int,
-        litre: Int,
+        litre: Double,
         expense: Int,
         lat: Double,
         long: Double
@@ -281,7 +281,7 @@ class VeeRepository(
         token: String,
         date: String,
         distance: Int,
-        litre: Int,
+        litre: Double,
         expense: Int,
         lat: Double,
         long: Double
