+++ app/src/main/java/id/vee/android/domain/repository/VeeDataSource.kt
@@ -29,7 +29,7 @@ interface VeeDataSource {
         token: String,
         date: String,
         distance: Int,
-        litre: Int,
+        litre: Double,
         expense: Int,
         lat: Double,
         long: Double
@@ -74,7 +74,7 @@ interface VeeDataSource {
         token: String,
         date: String,
         distance: Int,
-        litre: Int,
+        litre: Double,
         expense: Int,
         lat: Double,
         long: Double
